==> notebooks/09_LangModel_API_Demo.ipynb <==
This notebook shows you how ping the microservice that allows you to retrieve features given an issue title and body. 

import requests
import json
import numpy as np
from passlib.apps import custom_app_context as pwd_context
API_ENDPOINT = 'http://gh-issue-labeler.com/text'
API_KEY = 'YOUR_API_KEY' # Hamel can give you this

data = {'title': 'Fix the issue', 
        'body': 'I am encountering an error\n when trying to push the button.'}

# sending post request and saving response as response object 
r = requests.post(url=API_ENDPOINT,
                  headers={'Token':pwd_context.hash(API_KEY)},
                  json=data)

np.frombuffer(r.content, dtype='<f4')
